Provide multi-driver wrapper and test
An implementation of the multi_driver decorator is provided in
multi_device_support.py.
In order to test it the file i2c_device_mocks provides some simple
mocks for sensor, connection and driver.
test_multi_device shows the usage of the multi_driver decorator in both
sequential and concurrent mode.