Provide I2cMultiSensorBridgeConnection
With the I2cMultiSensorBridgeConnection several
SensorBridges can be combined into one virtual divce with many ports
channels.
The I2cMultiSensorBridgeConnection.get_multi_channel() is used to
create an AbstractMultiChannel. This MultiChannel has a channel for each
available port of the sensor bridge.
It can be used to communicate in simultanously with the connected sensors.