Support for array to integer conversion
Various sensors use 48bit integers as serial number. In the specifications
i defined these serialnumbers as uint8[6].
According to Roman Schilter this serial numbers are supposed to be
seen as integer and not as byte array.
Since it is a common situation that a byte array has to be interpreted
in a specific way a first converstion to int is added that can be
activated by configuration.
The generator will be adapted to support this configuration in the
specification.