Improve serialization or arrays and strings
The classes RxData and TxData are modified such that they allow a more
consistent usage of arrays and strings even in transfers that do not
have only one sequence argument.
Arrays can be passed as list or tuple and will be packed properly into a
byte-array. When receiving arrays they will be returned as one tuple
with the respective element type.
But: only one string may be present in the argument list of a transfer!